requirements.txt
Flask==3.0.3
boto3==1.34.99
python-dotenv==1.0.1